+++ tests/test_data_sets.R
@@ -2,7 +2,8 @@ stopifnotdataset = function(data) {
   stopifnot(
     is.data.frame(data),
     colnames(data) == tolower(colnames(data)),
-    colnames(data) == make.names(colnames(data), unique = TRUE)
+    colnames(data) == make.names(colnames(data), unique = TRUE),
+    is.integer(attr(data, "row.names"))
   )
 }
 
@@ -13,3 +14,14 @@ stopifnot(nrow(kc_housing) == 21613L, ncol(kc_housing) == 20L)
 data("titanic", package = "mlr3data")
 stopifnotdataset(titanic)
 stopifnot(nrow(titanic) == 1309L, ncol(titanic) == 11L)
+
+data("penguins", package = "mlr3data")
+stopifnotdataset(penguins)
+stopifnot(nrow(penguins) == 344L, ncol(titanic) == 8L)
+
+
+if (requireNamespace("mlr3")) {
+  stopifnot(inherits(mlr3::tsk("kc_housing"), "TaskRegr"))
+  stopifnot(inherits(mlr3::tsk("titanic"), "TaskClassif"))
+  stopifnot(inherits(mlr3::tsk("penguins"), "TaskClassif"))
+}