+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Data sets are now automatically added to the task dictionary if mlr3 is
   loaded.
+* Added penguins data from the `palmerpenguins` package.
 
 # mlr3data 0.1.0